requirements.txt
pylint
flake8
pyyaml
paramiko
